// part9/firstStepTypescriptReact/src/interfaces/courseParts.interface.ts
interface CoursePartBase {
  name: string;
  exerciseCount: number;
}
interface CourseDescription extends CoursePartBase{
   description: string;
}
interface CoursePartBasic extends CourseDescription {
  kind: "basic"
}

interface CoursePartGroup extends CoursePartBase {
  groupProjectCount: number;
  kind: "group"
}

interface CoursePartBackground extends CourseDescription {
  backgroundMaterial: string;
  kind: "background"
}
interface CoursePartSpecial extends CourseDescription{
  requirements: string[];
  kind: 'special'
}
export type CoursePart = CoursePartBasic | CoursePartGroup | CoursePartBackground | CoursePartSpecial;